pyproject.toml
[project]
name = "sph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["matplotlib>=3.10.1", "warp-lang>=1.7.0"]

[tool.basedpyright]
typeCheckingMode = "off"
reportAny = false
reportExplicitAny = false
reportInvalidTypeForm = false
reportUnknownParameterType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnusedCallResult = false
reportConstantRedefinition = false
